Satoshi to Pagella
Replaced corporate typeface Satoshi with TeX Gyre Pagella because while Satoshi was absolutely beautiful, it was cumbersome to use. Customizing it for the web is really easy with font-feature-settings but getting it to look the same in Word is impossible without breaking the terms of the license. Pagella is under a more permissive license so I can develop better cohesiveness going forward.